Abort on unsupported accesses to the CPU state
Currently we do not support access to pointer data types in the CPU state
structure. Now, since their usage is marginal, instead of failing at
compile-time we put an abort instruction when the code reaches that
point.

The same is done in case an access to an array member using a
non-constant index is performed.