Handle double translations due to VLE
In variable-length encoding architectures, sometimes you might have one
instruction containing another one. This is not a problem until you get
the next one, which would be translated twice, in fact, it's not at the
beginning of a basic block and the current mechanism to detect already
translated code does not handle it.

This patch makes `JumpTargetManager::newPC` check the whole map of the
translated instructions. This might have to be improved in the future.