Handle the PC += arg case
Some hand crafted assembly code perform a PC-relative jump of an
non-statically known amount. This patch introduces a simple hack to
handle such a situation by simply detecting it and marking as potential
jump targets all the instructions to come until the next jump.

This is implemented by the `JumpTargetManager::handleSumJump` and
`isSumJump` functions.

This commit also introduces a new implementation of `getNextPC` not
requiring the dominator tree.