Introduce `OnceQueue`
`OnceQueue` is a queue which not only keeps track of what's already in
the queue and prevents it from being re-inserted, but also keeps track
of what has ever been in the queue, and prevents it from returning to
the queue.

`OnceQueue` is implemented by adding a new template parameter to what
was once `UniquedQueue`.