Introduce `check-conventions.sh`, `.clang-format`
`.clang-format` is the configuration file for the `clang-format` tool,
which can help us to enforce a consistent coding style.

`check-conventions.sh` is a a simple bash script that checks (using
mostly `git grep`) if, after running `clang-format`, there are some
undesired situations such as lines ending with "(" or "<".

This script should produce no output before a merge request is
merged. However, currently this is not the case, therefore the script
should be used mostly for new code (in particular, files) only.